Added custom hooks in the Dashboard Page
